+++ Exiled.Events/Patches/Generic/GhostMode.cs
@@ -7,7 +7,7 @@
 
 namespace Exiled.Events.Patches.Generic
 {
-    #pragma warning disable SA1313
+#pragma warning disable SA1313
     using System;
     using System.Collections.Generic;
     using CustomPlayerEffects;
@@ -101,7 +101,7 @@ namespace Exiled.Events.Patches.Generic
                     else
                     {
                         byte part;
-                        for (part = (byte)0; (int)part < __instance.usedData / 20; ++part)
+                        for (part = 0; part < __instance.usedData / 20; ++part)
                             networkConnection.Send(new PlayerPositionManager.PositionMessage(__instance.transmitBuffer, 20, part), 1);
                         byte count = (byte)(__instance.usedData % (part * 20));
                         if (count > 0)
